Renaming unittests
